Improved calc_matrix_operators_grid speed
Greatly improved the speed of the calc_matrix_operators_grid routine, which was slowing down model initialisation a lot (as it's used during the calculation of grid-to-mesh mapping operators). It's got to do with the way elements are added to CSR matrices, where every time the whole ptr array needs to be updated. This could maybe be improved further in the future so that other model parts benefit as well. Basically it would mean making sure that a CSR_filling loop always goes over ALL the (local) rows in order, so that you could get away with only updating ptr( i+1) instead of the current ptr( i+1 : n). Anyway, that's for later.