Added single-variable NetCDF debug files, fixed small bugs in remapping and SSA solver
Fixed two tiny typos in the trace-line-through-Voronoi-cells code that could on rare occasions cause grid-to-mesh or mesh-to-mesh remapping to crash.

Fixed implementation of Neumann ("infinite") boundary conditions to the stress balance at the domain border. Using the d/dx, d/dy operator matrices sometimes causes artefacts in the velocity solution ("bumps" at the border). Using the old-style "average over neighbouring triangles" is a bit more diffusive, but much more stable.

Tested everything on the SSA_icestream experiment. It's difficult to reproduce the analytical solution, as the boundary conditions are a pain in the backside to get right (due to the high du/dy on the east/west borders) but if you use a very large domain, velocities in the centre of the domain approach the analytical solution.