Added the sans option for the SSA/DIVA solvers
Added the option to neglect the gradients of the effective viscosity (the "cross-terms") in the SSA/DIVA, leading to the "SSA_sans" or "DIVA_sans" approximations.

Note that there is no clear mathematical or physical reason why this should be allowed. However, while I have found a few cases where there are noticeable differences in the solutions (e.g. ISMIP-HOM experiments with high strain rates), most of the time the difference with respect to the full SSA/DIVA is very small. The "sans" option makes the solver quite a lot more stable and therefore faster. Someone really ought to perform some proper experiments to determine whether or not this should be the default.