Added ISMIP-HOM experiment A as an ice unit test
Added ISMIP-HOM A as well. Also expanded the Matlab scripts to generate the Pattyn2008-style transect plots.

The BPA results seem a bit too diffusive in experiment C. Not sure why this is, but since this only concerns small-scale variations in basal roughness (smaller than the basal inversion can produce anyway), it doesn't seem to be an urgent problem.