Added first calculation of dH/dt
Added the first routines to calculate ice thickness rates of change (dH/dt). This uses a time-explicit scheme, where a matrix operator is defined (based on the ice velocities) that can be multiplied with the ice thickness to yield the ice flux divergence. This same operator can (will) later be used to set up a time-implicit scheme as well.

Added the Halfar dome transient experiment as an ice unit test to confirm that this scheme works. As dynamic time stepping is not yet available it is not particularly fast, so it is only run for 5,000 years at 40 km resolution.