Added semi-implicit calculation of dHdt
That. Verified that it works with the Halfar dome experiment. It does generate a bunch of negative H warnings. I vaguely remember that in IMAU-ICE this happened at the ice margin, which I solved by just using the explicit scheme there (which also makes implementing stuff like calving a lot easier). Might do that again.