Small fix, unit tests work again
The two Halfar dome unit tests accidentally used the same name for their meshes, which caused problems with the remapping Atlas. That's fixed now.
Also made it so that the warnings for "negative ice thickness found in dH/dt calculation" are only triggered when that thickness is below -0.1 m. The (semi-)implicit sometimes produces VERY small negative thicknesses (e.g. -2e-189 m), which can be safely limited to zero.
Also also added comments with a nice mathematical derivation to the ice thickness solvers.