Rearranged sme modules
Rearranged some code over different subroutines and modules.

Made the ice flow law into a choice ('Glen' being the only option).

Wrote a nice Matlab script to make adding/reordering config parameters MUCH easier. Basically, it reads the first 1/4th of the model_configuration module to get all the _config parameters. Then it substitutes those (in the appropriate format) into the config_type definition, the namelist for reading config files, and the subroutine where the values from the _config parameters are copied to the config structure. This means you now can add the config parameters only once in the module, and then run the Matlab script to automatically add it in the other three places as well.

Also wrote a Matlab script for automatically generating the unit_tests config file from the model_configuration module, ensuring that all config parameters are present in the file.