Added flowline-based basal inversion
Added the basal inversion based on the flowline-averaged thickness error and thinning rates. Set this up in the MISMIP+ geometry (pretty much as described in Berends et al. 2023, but with a slightly different target bed roughness) and verified that it converges. Includes most of the stability enhancements Jorjo set up for Antarctica, so hopefully it will work there too.