Fix INTENT of predictor/corrector structure during restart
When initialising the predictor/corrector fields from a file
(e.g. during a restart), the predictor/corrector structure pc
was being passed as INTENT(OUT) to the file-reading routine,
which treated it as a new, unallocated structure. This triggered
an error due to the resulting 0-size of the pc fields. Fixed it
by declaring pc as INTENT(INOUT) in initialise_pc_scheme_from_file.