Add time step limit based on flux divergence div(Q)
New maximum limit to the largest possible time step based
on the flow of ice from vertex to vertex. Now the routines
that compute dH/dt have access to this limit, but so far
they don't make use of it (yet). Computed in subroutine
calc_flux_limited_timestep in ice_thickness.f90.