More stability improvements
Added automatic adapting of the regularisation parameter epsilon_sq_0 that goes into the effective strain rate for Glen's flow law, in case the viscosity iteration doesn't converge. Included a stop criterion for if the parameter becomes too big.

Added iteration of the P/C time step, following the Yelmo code. If, at the end of the PC timestep, the truncation error turns out to be too large, run it again with a smaller dt, until the truncation decreases to below the specified tolerance.