Various small fixes
Fixed a bug in mesh-to-mesh remapping that resulted in frequent errors in the remapped data field. The model now slows down much less after a mesh update as the remapped ice thickness field is much more accurate.

Slightly changed the automatic adaptation of the relaxation factor and effective strain rate regularisation parameter in the viscosity iterations of the SSA/DIVA/BPA to improve performance.

Added a calving front for the MISMIp geometry to make it look nicer (and make it suffer less from remapping problems at the domain border).

Added the truncation error of the P/C scheme as an optional output field.