Added out-of-time calving front relaxation
Relax ice thickness around the calving front. This routine "steps out of time" for a bit (default dt_relax = 2 yr), where it lets the ice thickness near the calving front relax for a little bit. To achieve this, it uses the BC_prescr_mask optional arguments of the velocity solver and the thickness solver to keep velocities and thickness fixed over parts of the domain. In this case, it keeps them fixed everywhere except over the open ocean, and the first 5 shelf vertices inward from the calving front. This allows the little bumps that appear in the ice thickness after remapping, which seriously slow down the velocity solution, to relax a little.