Refix parallel out-of-bounds: replace_Ti_with_robin_solution
Input argument Ti is used in the 3D thermo solver to receive
a local Ti_tplusdt, which was broken by the previous fix. The
other option is to declare the dimensions of the input Ti as
DIMENSION(mesh%vi1:mesh%vi2,mesh%nz), which looks awful but
works always. Well.