Add realistic ocean I/O and mesh remapping
Modified (or added rather added) NetCDF routines to
deal with the ocean external data I/O, which have
the particularity of not using the scaled zeta vertical
grid, but its own "depth" vertical coordinate. Thus,
existing routines (most of which check for zeta) could
not be used.

Reading of ocean data includes the mapping between the
original vertical layers of the data onto the model's
ocean vertical grid, and a simple extrapolation of
valid data into NaN areas (if a more sophisticated
method is needed, the user is expected to do this
themselves during pre-processing, and provide a good
looking data set from the start).

Added as well the creation and writing of the ocean
model's restart files, already on a region-dependent
manner.