Add zealous usage of no-ice mask + general minimum thickness
Added calls to the computation and application of the no-ice
mask after each modification to Hi and before main each computation
of other geometry-related quantities. This now includes an optional
absolute threshold for the minimum thickness allowed in the model,
under which thinner ice will be added to the no-ice mask and thus
removed automatically each time step.