Add reduced sliding for cold bases + min beta for thin ice (ZI-only)
All this only added for the Zoet-Iverson sliding law:

The basal temp w.r.t. to the pressure melt point is used to
scale the friction based on some temperature thresholds and
a linear ramp (values still hardcoded). The colder the base,
the less impact the till friction angle has on the total
effective pressure (and thus the till yield stress).

Also, added a linear increase in the minimum value allow
for the basal friction coefficient that depends on how
thin the ice is, below a certain threshold (hardcoded).
Idea is that very thin ice doesn't get crazy fast if the
till friction angle is small and/or if the surface slope
(and thus the driving stress is ultra big). This helps
stabilise the velocities over ice shelves, which otherwise
go nuts.