Add mods to flowline friction inversion
Replaced check for grounded vertices by a requirement of
fraction_gr being equal to 1 (i.e. fully grounded).

Skip adjustment in vertices that are already improving.

Add safety net during tracing of flowlines: if the distance
from origin of the next trace is smaller than the previous
one, stop the trace. It is very likely that it crossed a
divide (resolution overshoot) and it is now jumping from
one side of the divide to the other forever and ever.