Add Pien 2023 sliding inversion method (Tim's CISM)
Added a simple version of the inversion of sliding (or
rather the till friction angle) based on the method used
by Tim in CISM, but only including the first 2 regularisations
(misfit and dHdt; missing the relaxation target field). Added
this in the main module for now. Will move it later somewhere
else.