Add Ti_hom and pore_water_likelihood as model variables
Now the basal temperature relative to the pressure melting point
of ice is part of the ice model type, and is computed in its own
subroutine within the thermodynamics model. Also took the chance
to add a new variable, which stores the computed likelihood of
having subglacial water under the ice sheet, which will be used
later.