Fix typo in bedrock CDF + added threshold calving to no-ice mask
Routine calc_grounded_fractions_bedrock_CDF_b was being called with
fraction_gr_CDF_a as the output variable (note that the latter is
the variable on the a-grid).

Also added the "threshold" calving approach as an option in the
no-ice mask. This way, ice is not allowed to survive when it is
thinner than X m, effectively removing it.