Update PC method to accomodate BMB inversion
Improve consistency in the computation of the predicted/corrected
geometries after modifying the ice thickness evolution and to
enable an easy inversion of BMB under floating ice. BMB is now
always updated before the corrector step, to make sure that the
BMB field is consistent with the predicted mask and geometry.