Make inverted BMB take in full target dHi_dt
Previously it was only accounting for the negative target
thinning rates over ice shelf areas (Smith 2019 looks weird
in those areas). That created an inconsistency with the final
computed dHi_dt. Fixed it by simply taking all of it. That's a
problem in the input data, not in the model.