Add other MB sources to computation of integrated fluxes
Additional mass sources, e.g., target thinning rates, were
not accounted for (or rather not reported) in the scalars.
This made, e.g., the total outflux from the ice sheet to not
match the total influx of mass, among other slightly annoying
things. Also fixed a miss-parallelisation issue in the same
routines (lack of MPI_ALLREDUCE for SMB and BMB).