Add detailed mass balance tracking
Vertical fluxes (SMB and BMB) are now tracked separately
for grounded, floating, ice-free land, and ice-free ocean
areas. In addition, imposed modifications of mass in the
shape of no-ice masks, negative thickness prevention, etc.,
are also tracked in the same way. This allows to close the
mass budget of the entire domain. Lateral fluxes through
transition zones are also recorded. All this is now part
of the scalar output file.