Added th hybrid DIVA/BPA solver
It works in ISMIP-HOM experiment B. Almost all the physics are used from the existing DIVA/BPA routines, so it should work pretty much anywhere.

To be possibly added in the future: determining the solver mask (where to DIVA and where to BPA) based on geometry/strain rates/something, to automatically detect where this is beneficial. Another option could be to provide a mask from an eternal file. But for now, just using the Regions of Interest seems a good start.