Lay groundwork for sub-grid margin fractions
Including the computation of the effective thickness at calving
fronts. Declaration, allocation, and remapping of both Hi_eff
and fraction_margin (ex-fraction_cf). The latter is already
part of the input arguments for calc_dHi_dt, and used to prevent
flux divergence from vertices whose are is not fully covered with
ice. This changes the partitioning between BMB and LMB for those
vertices, somthing that is at the moment WIP.