Make semi- and implicit Hi solvers flux-conserving
Within the solvers, the mass balance was capped at the
amount of available mass per vertex. While preventing
negative thickness, this had the colateral of disregarding
mass balance terms that have the potential of removing more
mass than available. A good example is basal melt in the
ocean. Another problem was that during that capping, the
flow converged into these vertices was not taken into
account, underestimating the available mass of a vertex.
Fixed both issues by tracking the disregarded change rates,
and also adding the divergence to the capping step.