Fixed the namelist file conflict
Fixed a problem where starting multiple runs on snellius simultaneously could lead to crashes. This was because the model creates a temporary text file of the config namelist, and havingmultiple simulations doing so together led to conflicts as they all tried to read/write the same file.
Fixed this by making the filename dependent on the config filename, so it should be different for each run.