Update equilibrium routines: MB_inversion + modiness
Reordered the calls to the MB inversion and ice thickness
modification routines within the PC method. Now these are
applied to both the predicted and corrected geometries and
thinning rates, improving the stability of the predicted
velocities.

Also, reintegrated modiness into limitness in the FML method,
as it created buggy behaviour during the linear release.