Add optional temperature BC at grounding line: subgrid, pmp
In addition to the existing one (heat flux at the base, which
assumes the cell is fully grounded) now the model can compute a
new temperature profile that combines the solutions for a fully
grounded and fully floating cell (latter using the pressure
melting point temperature as BC). This combination depends on
the fraction of the cell that is grounded/floating. In addition,
another option that assumes the grounding line is always at the
pressure melting point has been added as well. Again, this is
done only at the grounding line (grounded side).