Add no-ice Ellesmere, guiltiness, migrate effective thickness
For the Greenland domain, now the Ellesmere area can be completely
removed from the domain (including submerging the land so it does
not influence mesh generation).

The computation of the effective thickness has been moved to the
ice utilities module, since it was needed at some lines within the
alter_ice_thickness subroutine.

A new field has been added, which tracks the number of times each
vertex triggers an additional Predictor/Corrector iteration. This
field will be later used to regularise the inversion of sub-glacial
conditions.