[WIP] Update hydro inversion: time decay, guilty, extrapolations
Added a small time-dependent increase of the inversion time
scale, so the magnitude of the adjustment decreases over time
during the last part of an inversion (hard-coded for now).

Worked on the extrapolation of the inverted variable, giving
some thought to the use of grounded neighbours when assigning
inverted fields to the ice margins.

Added a new regularisation based on the number of times a
vertex has caused an additional iteration of the PC method
during the computation of the new ice thickness. The more times
a vertex hast triggered this reduction in the time-step, the
smaller the change in the inverted hydro field.