Clean up code: hydro inv + ice thickness
Made sure routine calls are consistent, and small
changes here and there to make things more robust.
Now the inversion of pore water fraction is called
from the main model subroutine, whereas the actual
computation of the pore water pressure is done in
within the call to the basal hydrology model. This
kinda mimics the structure for the inversion of the
bedrock roughness.