Add a few ROIs to the model + trimmer ROI checks
Added a couple of Antarctic locations: Amery and Riiser-Larsen
ice shelves in Antarctica, and the glaciers north of Narsarsuaq
in Greenland. Also, streamlined a bit the checks for the existence
of the requested ROIs within the code, to avoid having to add the
entire list of existing options to each model region every time.