Update extrapolation of pore water fraction and till yield stress
Removed the Gaussian extrapolation over non-inverted areas, and
replaced it by a neighbour search: first at the grounded margins,
it looks for the neighbour with the highest fraction and copies
its value. Then, the same for the floating side of the grounding
line. Then, the same for ice-free land. This inverted field is
then passed on to the main run hydrology routine, which computes
the pore water pressure based on the pore water fraction and the
grounded area fractions, to obtain a smooth field that maximises
pore water pressure over floating areas even without doing an
inversion. A similar thing is done for the till yield stress over
ice-free land areas (removed the Gaussian extrapolation there as
well).

Added a new method "none" for hydrology, which assumes a zero pore
water fraction, except over floating, ocean, and partially floating
areas.