[WIP] Update reconstruction of SMB method: MB inversion and parameters
Now the reconstruction method for SMB defines a number of free
parameters: equilibrium/transitional/desertification line
altitudes, max and min SMB, and a smoothing radius. These are
still hard-coded, and work only for Patagonia.

The SMB reconstruction now is also somewhat embedded in the
MB inversion routines, preventing the inversion within the
polygon that defines the reconstructed limits of the former
ice sheet. Again, this only works for Patagonia for now.