Add geometry relaxation
A routine that's called during initialisation, which
runs the ice dynamics model in an out-of-time fashion
to let the initial geometry flow a bit undisturbed by
inversions or mass balance terms to relax its noisy
observational data origins. The amount of pseudo years
this routine runs can be set in the config files. The
relaxed geometry is then written to the reference fields
as well, to prevent the inversions targeting the unrelaxed
data.