[WIP] Revamp basal inversion and computation of basal friction
The inversion of pore water fractions now can use target
surface ice velocities to nudge the modelled ice thickness.
The inverted field is then extrapolated based on traced flow
lines downstream of the inverted area. The extrapolated field
is then combined with a new term that represents ocean entrainment
beneath the ice. This combination makes use of the grounded area
fractions, which aims to replace their use later in the code
within the velocity modules (optional). The final pore water
fraction field + ocean entrainment is then used to compute the
effective pressure.

Similarly, the bed roughness module now computes a quantity
called bed roughness (duh), which includes the tangent of the
till friction angle and the grounded area fractions.

Both fields are recomputed every time step to account for changes
in sea level and grounded area fractions (as ice thins or thickens).
Both fields are then multiplied to obtain the till yield stress in
the Zoet-Iverson sliding law. Other sliding laws will follow soon.

In summary, the multiplication of the basal friction field and the
grounded area fraction now can take place within the basal modules,
insted of the velocity modules, which provides a much finer control
over where it is applied. This helps prevent numerical instabilities
in the velocity field that are common in realistic setups.