Add independent scaling of basal conditions on the A grid
If so desired, the pore water fraction and bed roughness fields
can now be scaled based on the grounded area fractions on the A
grid, before the basal friction coefficient is computed. This
allows for a finer control of the final field, which is particularly
useful in realistic experiments with very complex geometries.
The old scaling of beta on the B grid within the velocity modules
is still the default, for which its exponent can be set to 0 in
the configuration files if the new scaling is preferred.