Fix bug in mesh refinement: sheet polygon inside shelf polygon
For rare geometries (e.g. the donut-shaped CalvingMIP one), the
polygon that represents the ice shelves is not a closed one (thus
the "donut"). In those cases, the ice sheet polygon lies perfectly
inside the ice shelf one, which made the refinement overwrite
the ice sheet config parameters by the floating ones. Fixed it by
skipping the shelf refinement when a point lies both inside a shelf
and a sheet polygon at the same time. This is done through a new
optional input argument in the routines, which contains a "no-refinement"
zone, in this case described by the ice sheet polygon.