Lay groundwork for formal lateral and artificial mass balance
In order for LMB and AMB to be part of the continuity equation,
they need to be leveled with the other mass balance terms. Added
the missing LMB and AMB modules, routines, and variables that
make these 2 terms similar to the other ones. Also, checked that
these can be remapped after a mesh update. Now the AMB term
contains a good chunk of what the old dHi_dt_residual variable
used to track. Idea is that this structure contains fields that
describe the amount of ice thickness modifications that each
big area of the code imposed.

Also updated the computation of effective ice thickness to allow
ice shelf point to attain the thickness of grounded neighbours as
well, which helps with the "inception" of floating ice.

Also updated the MB inversion to store the inverted values in a
sub-field of their respective MB data structures.