Reorder some types
Reorder the grid and grid_lonlat types in the makefile (and correspionding include statements) to make sure we can define all the derived types right at the start of the compilation. Needed to do this so I can start building the "test_eq_grid", test_eq_grid_lonlat" and "test_eq_mesh" routines.