Add mapping/discretisation component tests
Added component tests for the mesh matrix operators for mapping/derivatives between the a/b/c-grids. These are automatically run on all the test meshes created by the mesh creation component tests.
Also added a Matlab script to analyse the results, creating for each mesh/test function combo the set of multipanel plots showing the results, and writing the maximum error for each matrix operator to a text file (which can eventually be used as en entry into the scoreboard).