Add grid-to-mesh remapping component tests
Use the suite of auto-generated test meshes (and x/y-grids, which I added too) to test the grid-to-mesh remapping code. It does this on all possible grid/mesh-combinations, which currently takes about 1 minute. As with the discretisation stuff, it writes the resulting fields to NetCDF files, which are analysed by a Matlab script, resulting in a little test report.

NOTE: a bunch of the tests are skipped because they crash on bugs in the remappng code. I will fix this later - and also, this is exactly why we need these tests!