Rename mesh/Delaunay modules
No need to repeat the (sub)directory names in the module name.

Note that it is not possible to have a routine named 'henk' in a module that is also named 'henk'. So e.g. the module is now named 'split_edges' (with a plural 's' at the end) so that the routine can remain 'split_edge'.