Refactor assertions/unit tests
There now exist a single "assert" and a single "unit_test" subroutine, both of which take a logical scalar as their input. Since all the basic test functions are now pure functions returning a logical scalar, they can be passed directly as the input argument for assert and unit_test, simplifying the code.