Indicate that earlier unit tests failed
With the increasing number of unit tests, it becomes annoying having to scroll up the terminal to check if all tests passed. Now, when one test fails, all subsequent passing tests are shown in yellow rather than green, so that the last message will be green if and only if all tests passed.