Update ice_model_main.f90
Small bugfix. With the standard compiler settings, floating-point variables (like t_pseudo) are initialised as NaN. Here, t_pseudo was initialised to zero, but only on the master; the other processes did not explicitly initialise it, so it was set to NaN, which results in really weird errors down the line. And, of course, when not explicitly specifying these compiler settings, most compilers will initialise floating-points to zero, so the problem goes unnoticed.