Add raw particle NetCDF output
Added code for output raw particle data to NetCDF.

Also removed the 'flexible memory' for particles. Instead, a fixed (large) amount of memory is allocated during initialisation, and if the number of particles exceeds that, the model crashes. While this might sometimes cause a bit more user frustration, it makes it infinitely easier for me to create NetCDF output.