Remove temporary scoreboard files
Temporary scoreboard files no longer exist. Instead, a separate file is created for each experiment+commit combination, with the git hash string included in the filename. This prevents merge conflicts when multiple pull requests are open simultaneously. The downside is that the number of individual scoreboard files will become very large, but since they are proportionally smaller, the total amount of disk space required should not be affected (and will still be very small, as they are very short text files).