Determine and output mask_ROI
Created mesh%poly_ROI which saves polygon on region of interest (ROI) such that you can use that later to determine the mask of the ROI. Created new variable ice%mask_ROI to store this mask and also to output. Current commit is working on 1 core and for 1 region of interest.