Add routines laddie_ROI
Now possible to select laddie_py as BMB model for the ROI. As input, it needs the BMB field for the entire grid (!).
Laddie runs using the geometry in the regional gridded output only, but it than outputs a file (now called BMB_latest_v00.nc) which has the BMB data on the full domain (zeros out of ROI). It was not possible to read in BMB_latest for the ROI because some Voronoi cells don't have data and then it crashes in the mesh/remapping/create_maps_grid_mesh module. Making adjustments in this module might be more efficient.