Fix subglacial discharge
Now it compiles but also runs. Still need to polish some stuff, including the output of subglacial discharge that is added to the mixed layer (laddie%SGD in m/s). Need to check whether it also works properly if setting C%choice_laddie_SGD to 'none'. Also now the amount is only added to the thickness of the layer, but no change is implemented on the T,S of the layer.