Improve BMB inversion at calving front
Since the remapped ice thickness (of the target geometry) at the calving front is too thin (because of the problem of remapping a discontinuous field), and the ice dynamics wants to make it thicker, the BMB inversion is compensating ror this by applying a lot of 'artificial' melt at the last floating grid cell. Fixed this by defining the target thickness there as the mean over the non-front shelf neighbours. Doesn't fix things completely, but it does get a lot better.