Refactor the CMake code
UPSY is now compiled as a static library, to which the executables are linked. This prevents conflicts and speeds up the compilation.

Moved the linking to external dependencies (OpenMPI, PETSc, NetCDF), and the application of compiler flags, to CMake functions to prevent code duplication.

Right now, two executables are compiled: UPSY_unit_test_program, and UFEMISM_program. More executables can be easily defined in CMakeLists.txt.

To do: provide input argumnts to CMakeLists.txt to tell it which executables should be compiled, so that e.g. running ./compile_LADDIE will not compile UFEMISM.